+++ src/lib/EntityRelationTypes.ts
@@ -1,18 +1,23 @@
 import CoreEntity from '../classes/CoreEntity';
-import { EntityConfig } from '../classes/annotation';
+import {
+  EntityConfig,
+  EProperties,
+  EUpDateProperties,
+} from '../classes/annotation';
 
 export interface ICoreEntityHandler {
   createEntity<E extends CoreEntity>(
     config: EntityConfig<E>,
-    entity: E
-  ): Promise<E | null>;
+    entity: EProperties<E>
+  ): Promise<E>;
   updateEntity<E extends CoreEntity>(
     config: EntityConfig<E>,
-    entity: E
-  ): Promise<E | null>;
+    e_id: number,
+    entity: EUpDateProperties<E>
+  ): Promise<boolean>;
   getEntityById<E extends CoreEntity>(
     config: EntityConfig<E>,
-    id: number
+    e_id: number
   ): Promise<E | null>;
   deleteEntityById(className: string, id: number): Promise<boolean>;
   getEntityList<E extends CoreEntity>(