+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.15.1] - 2021-12-06
+### Added
+- CoreError # log to error log chanel
+### Changed
+- Interface for entity update 
+- Better interfaces for entity constructor
+- Update InMemDb
+
 ## [0.15.0] - 2021-12-02
 ### Added
 - Add Changelog